requirements.txt
lean-dojo==4.20.0
mistralai==2.1.3
python-dotenv==1.2.2
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
z3-solver>=4.13.0
numpy>=1.26.0
sentence-transformers>=3.0.0
datasets>=2.20.0
pytest>=8.0.0
pylatexenc>=2.10
